backend/requirements.txt
websockets==12.0
opuslib==3.0.1
numpy==1.26.4
python-dotenv==1.0.0